requirements.txt
flask
python-dotenv
gunicorn
sendgrid>=6.0.0    

flask-cors